Refactor: Footer 이름 BottomNavbar로 변경
