+++ src/app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 import "./globals.css";
-import Footer from "@/components/common/Footer";
+import Footer from "@/components/common/BottomNavbar";
 
 export const metadata: Metadata = {
   title: "Netflix",
